+++ src/apply_to_buttons.js
@@ -1,9 +1,9 @@
 import { log } from './utils.js';
-const apply_to_buttons = function( opt= {} ){
-    opt = Object.assign({}, { color: '', debug: false }, opt);
-    opt.color = opt.color[0].toUpperCase() + opt.color.substring(1, opt.color.length );
-    const COLOR_CHAR = opt.color[0].toUpperCase();
-    const CLASS_STR = 'btn ' + opt.color + '-tag btn-lg';
+const apply_to_buttons = function( COLOR = {} ){
+    COLOR = Object.assign({}, { color: '', debug: false }, COLOR);
+    COLOR.color = COLOR.color[0].toUpperCase() + COLOR.color.substring(1, COLOR.color.length );
+    const COLOR_CHAR = COLOR.color[0].toUpperCase();
+    const CLASS_STR = 'btn ' + COLOR.color + '-tag btn-lg';
     const buttons = document.getElementsByTagName('button');
     let i = 0, len = buttons.length;
     while(i < len){
@@ -12,7 +12,7 @@ const apply_to_buttons = function( opt= {} ){
         if(arr_id[0] != 'W' && arr_id[0] != 'U' && arr_id.length === 5){
              el.id = COLOR_CHAR + arr_id.slice(1, 5).join('');
              el.className = CLASS_STR;
-             if(opt.debug){
+             if(COLOR.debug){
                  log( 'id='+ el.id, 'className=' + el.className );
              }
         }